renamed plots/graphs.py to plots/graphs_demand.py
