+++ cea/plots/graphs_demand.py
